src/quantum_meta/meta_cognitive/processor.py
"""
Quantum Meta-Cognitive Framework with Numerical Stability
------------------------------------------------------
Implements rigorous quantum state processing with explicit
floating-point precision control.

Mathematical Framework:
Γ = {H, Ψ, M, ε}

where:
H: Hilbert space
Ψ: Wave function
M: Measurement operators
ε: Numerical precision threshold
"""

import numpy as np
from typing import Dict, List, Tuple, Optional
from dataclasses import dataclass
from pathlib import Path
import json
from scipy.linalg import sqrtm

# Numerical precision constants
EPSILON = 1e-12  # Floating point precision threshold
MAX_COHERENCE = 1.0 - EPSILON  # Maximum allowed coherence value

@dataclass
class MetaState:
    """
    Quantum-mechanical state representation with bounded coherence
    
    Constraints:
    0 ≤ coherence ≤ 1
    """
    patterns: np.ndarray
    coherence: float
    emergence: Dict[str, float]

class PatternRecognitionManifold:
    """
    Quantum pattern recognition with numerical stability
    
    Core Operations:
    1. Numerically stable state normalization
    2. Bounded coherence computation
    3. Precision-aware pattern extraction
    """
    
    def __init__(self, dimension: Tuple[int, ...]):
        self.dimension = dimension
        self.recognition_field = np.zeros(dimension, dtype=complex)
        
    def _normalize_quantum_state(self, state: np.ndarray) -> np.ndarray:
        """
        Normalize quantum state with numerical stability
        
        |ψ_norm⟩ = |ψ⟩/√(⟨ψ|ψ⟩ + ε)
        """
        norm = np.sqrt(np.abs(np.vdot(state, state)) + EPSILON)
        return state / norm
    
    def _bound_coherence(self, value: float) -> float:
        """
        Enforce coherence bounds with numerical stability
        
        C_bounded = min(max(C, 0), 1-ε)
        """
        return np.clip(value, 0.0, MAX_COHERENCE)
    
    def _compute_quantum_kernel(self, x: complex, y: np.ndarray) -> complex:
        """
        Numerically stable quantum kernel
        
        K(x,y) = ⟨x|y⟩/√(⟨x|x⟩⟨y|y⟩ + ε)
        """
        x_norm = np.abs(x) + EPSILON
        y_norm = np.sqrt(np.abs(np.vdot(y.flatten(), y.flatten())) + EPSILON)
        
        kernel_value = np.mean(np.conj(x) * y.flatten()) / (x_norm * y_norm)
        return kernel_value

    def recognize_patterns(self, state: np.ndarray) -> np.ndarray:
        """
        Extract quantum patterns with numerical stability
        """
        normalized_state = self._normalize_quantum_state(state)
        field = np.zeros_like(state, dtype=complex)
        
        for idx in np.ndindex(state.shape):
            field[idx] = self._compute_quantum_kernel(
                normalized_state[idx],
                normalized_state
            )
            
        return self._normalize_quantum_state(field)

class MetaCognitiveProcessor:
    """
    Quantum-aware meta-cognitive processor with numerical stability
    """
    
    def __init__(self, dimension: Tuple[int, ...], quantum_numbers: Tuple[int, int, int]):
        self.dimension = dimension
        self.quantum_numbers = quantum_numbers
        self.pattern_recognizer = PatternRecognitionManifold(dimension)
        self.state_path = Path("meta_state.json")
        self._initialize_state()
        
    def _initialize_state(self):
        """Initialize quantum state tracking"""
        if self.state_path.exists():
            with open(self.state_path, 'r') as f:
                self.state = json.load(f)
        else:
            self.state = {
                'processed_states': 0,
                'average_coherence': 0.0,
                'pattern_history': []
            }
            
    def _compute_quantum_coherence(self, state: np.ndarray) -> float:
        """
        Compute bounded quantum coherence
        
        C(ρ) = min(|⟨ψ|ψ⟩|, 1-ε)
        """
        normalized_state = self.pattern_recognizer._normalize_quantum_state(state)
        coherence = float(np.abs(np.vdot(normalized_state, normalized_state)))
        return self.pattern_recognizer._bound_coherence(coherence)
    
    def process_quantum_state(self, state: np.ndarray) -> MetaState:
        """
        Process quantum state with bounded coherence
        """
        patterns = self.pattern_recognizer.recognize_patterns(state)
        coherence = self._compute_quantum_coherence(state)
        emergence = self._analyze_emergence(patterns)
        
        self._update_history(coherence, emergence)
        
        return MetaState(patterns, coherence, emergence)
    
    def _analyze_emergence(self, patterns: np.ndarray) -> Dict[str, float]:
        """
        Analyze quantum pattern emergence with bounded metrics
        """
        return {
            'complexity': self.pattern_recognizer._bound_coherence(float(np.abs(patterns).std())),
            'stability': self.pattern_recognizer._bound_coherence(float(np.real(patterns).mean())),
            'coherence': self._compute_quantum_coherence(patterns)
        }
        
    def _update_history(self, coherence: float, emergence: Dict[str, float]):
        """Update quantum state history with bounded values"""
        self.state['processed_states'] += 1
        self.state['average_coherence'] = self.pattern_recognizer._bound_coherence(
            (self.state['average_coherence'] * (self.state['processed_states'] - 1) +
             coherence) / self.state['processed_states']
        )
        
        self.state['pattern_history'].append({
            'timestamp': str(np.datetime64('now')),
            'coherence': float(coherence),
            'emergence_metrics': emergence
        })
        
        with open(self.state_path, 'w') as f:
            json.dump(self.state, f)