tests/quantum_meta/meta_cognitive/test_processor.py
"""
Meta-Cognitive Processor Test Suite
--------------------------------
Validates meta-cognitive processing system with quantum integration.

Test Framework:
Γ_test = {Ψ_test, M_test, E_test}

where:
Ψ_test: Quantum state tests
M_test: Meta-cognitive tests
E_test: Emergence tests
"""

import pytest
import numpy as np
from pathlib import Path
from src.quantum_meta.meta_cognitive.processor import (
    MetaCognitiveProcessor,
    MetaState,
    PatternRecognitionManifold
)

class TestMetaCognitiveProcessor:
    """
    Comprehensive test suite for meta-cognitive processing
    """
    
    @pytest.fixture
    def processor(self):
        """Initialize test processor"""
        return MetaCognitiveProcessor(
            dimension=(10, 10, 10),
            quantum_numbers=(1, 0, 0)
        )
    
    @pytest.fixture
    def quantum_state(self):
        """Generate test quantum state"""
        return np.random.randn(10, 10, 10) + 1j * np.random.randn(10, 10, 10)
    
    def test_pattern_recognition(self, processor, quantum_state):
        """
        Test pattern recognition system
        
        Validation:
        1. Pattern dimensionality
        2. Coherence preservation
        3. Recognition stability
        """
        patterns = processor.pattern_recognizer.recognize_patterns(quantum_state)
        
        # Validate dimensions
        assert patterns.shape == quantum_state.shape
        
        # Check coherence
        coherence = np.abs(np.vdot(patterns, patterns))
        assert 0.0 <= coherence <= 1.0 + 1e-10
        
        # Verify stability
        patterns2 = processor.pattern_recognizer.recognize_patterns(quantum_state)
        assert np.allclose(patterns, patterns2)
    
    def test_state_persistence(self, processor, quantum_state):
        """
        Test state persistence
        
        Validation:
        1. State saving
        2. State loading
        3. History tracking
        """
        # Process state
        meta_state = processor.process_quantum_state(quantum_state)
        
        # Verify state file exists
        assert Path("meta_state.json").exists()
        
        # Check state contents
        assert isinstance(meta_state, MetaState)
        assert meta_state.patterns.shape == quantum_state.shape
        assert 0.0 <= meta_state.coherence <= 1.0
        assert all(k in meta_state.emergence for k in 
                  ['complexity', 'stability', 'coherence'])
    
    def test_emergence_analysis(self, processor, quantum_state):
        """
        Test emergence analysis
        
        Validation:
        1. Metric computation
        2. Temporal evolution
        3. Pattern stability
        """
        # Initial state
        state1 = processor.process_quantum_state(quantum_state)
        
        # Evolved state
        evolved_state = quantum_state * np.exp(1j * 0.1)
        state2 = processor.process_quantum_state(evolved_state)
        
        # Verify emergence metrics
        assert state1.emergence.keys() == state2.emergence.keys()
        assert all(isinstance(v, float) for v in state1.emergence.values())
        
        # Check temporal evolution
        assert len(processor.state['pattern_history']) >= 2
    
    def test_quantum_compatibility(self, processor, quantum_state):
        """
        Test quantum state compatibility
        
        Validation:
        1. Phase preservation
        2. Superposition handling
        3. Measurement consistency
        """
        # Process quantum state
        meta_state = processor.process_quantum_state(quantum_state)
        
        # Verify phase information
        assert np.iscomplexobj(meta_state.patterns)
        
        # Check superposition
        superposition = (quantum_state + np.roll(quantum_state, 1)) / np.sqrt(2)
        super_state = processor.process_quantum_state(superposition)
        
        # Validate measurement
        assert abs(super_state.coherence - 1.0) < 1e-10